cleared one header. Moved buttons to header
