+++ rflib/main/default/lwc/rflibLogEventMonitor/rflibLogEventMonitor.css
@@ -31,14 +31,10 @@
     align-items: stretch;
 }
 
-.left-column {
-    transition: width 0.3s ease;
-}
-
-.right-column {
-    transition: width 0.3s ease;
-}
-
 .full-width {
     width: 100%;
+}
+
+.title {
+    font-size: 1.5rem;
 }
