+++ rflib/main/default/lwc/rflibLogEventViewer/rflibLogEventViewer.css
@@ -28,12 +28,6 @@
     color: #333;
 }
 
-.field-settings-panel {
-    background-color: #f8f9fa;
-    border: 1px solid #d0d0d0;
-    border-radius: 0.25rem;
-}
-
 .json-message {
     background-color: #f8f9fa;
     border: 1px solid #dee2e6;
