+++ rflib/main/default/staticresources/rflib_HidePageHeader.css
@@ -1,0 +1,3 @@
+app_flexipage-header {
+    display: none !important;
+}
